pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dtctl"
version = "0.1.0"
description = "A kubectl-inspired CLI for managing Dynatrace platform resources"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["dynatrace", "cli", "devops", "monitoring", "observability"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
    "Topic :: Utilities",
]
dependencies = [
    "typer[all]>=0.9.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "platformdirs>=4.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23",
    "mypy>=1.8",
    "ruff>=0.2",
    "pre-commit>=3.6",
]

[project.scripts]
dtctl = "dtctl.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/python-dtctl"
Documentation = "https://github.com/yourusername/python-dtctl#readme"
Repository = "https://github.com/yourusername/python-dtctl"
Issues = "https://github.com/yourusername/python-dtctl/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/dtctl"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["dtctl"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
